update region creation/deletion
