+++ vue/src/views/SAM.vue
@@ -3,6 +3,7 @@ import { onMounted } from "vue";
 import { ApiService } from "../client";
 import SAMEditor from "../components/imageViewer/SAMEditor.vue";
 import { ref } from 'vue';
+import { updateRegionList, updateRegionMap } from "../store";
 
 
 const props = defineProps(
@@ -18,6 +19,8 @@ const imageURL = ref('');
 const embeddingURL = ref('');
 
 onMounted(async () => {
+  await updateRegionList();
+  await updateRegionMap();
   const image = await ApiService.getSiteImage(parseInt(props.id, 10));
   imageURL.value = image.image;
   if (image.image_embedding) {
